Make outputs Long, because many problems need a long

+++ src/Template.kt
@@ -1,20 +1,20 @@
 fun main() {
     val day = "Day__DAY__"
 
-    fun part1(input: List<String>): Int {
+    fun part1(input: List<String>): Long {
         return -1
     }
 
-    fun part2(input: List<String>): Int {
+    fun part2(input: List<String>): Long {
         return -1
     }
 
     val testInput = readInput("${day}_test")
-    part1(testInput).assertEqual(-42)
+    part1(testInput).assertEqual(-42L)
 
     val input = readInput(day)
     timeAndPrint { part1(input) }
 
-    part2(testInput).assertEqual(-43)
+    part2(testInput).assertEqual(-43L)
     timeAndPrint { part2(input) }
 }
